requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.18
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.11.20
django-bower==5.2.0
django-celery==3.2.2
django-nvd3==0.9.7
docutils==0.14
gunicorn==19.6.0
holidays==0.9.9
Jinja2==2.10
kombu==3.0.37
MarkupSafe==1.1.0
numpy==1.16.1
pandas==0.24.1
Platypus-Opt==1.0.3
psycopg2==2.7.4
python-dateutil==2.8.0
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2018.9
redis==2.10.3
six==1.12.0
static3==0.7.0
statistics==1.0.3.5
Unidecode==1.0.23
vine==1.2.0
whitenoise==3.2.2
